Implement class Backend
