+++ backend/exception/ResumeNotFoundException.java
@@ -1,0 +1,12 @@
+package backend.exception;
+
+public class ResumeNotFoundException extends BackendException {
+
+    private static final String RESUME_NOT_FOUND_MESSAGE =
+            "Could not find resume!";
+
+    public ResumeNotFoundException() {
+        super(RESUME_NOT_FOUND_MESSAGE);
+    }
+
+}
