+++ backend/logic/ResumeGenerator.java
@@ -1,0 +1,11 @@
+package backend.logic;
+
+import backend.model.achievement.AchievementList;
+import backend.model.resume.ResumeImpl;
+
+public interface ResumeGenerator {
+    boolean isValidIndex(int index);
+    void selectAchievement(int index);
+    void deselectAchievement(int index);
+    ResumeImpl generateResume(AchievementList achievements);
+}
