+++ backend/model/achievement/Achievement.java
@@ -1,0 +1,6 @@
+package backend.model.achievement;
+
+public interface Achievement {
+    AchievementName getName();
+    AchievementContents getContents();
+}
