+++ backend/model/person/Person.java
@@ -1,0 +1,17 @@
+package backend.model.person;
+
+import backend.model.achievement.Achievement;
+import backend.model.achievement.AchievementList;
+import backend.model.resume.Resume;
+import backend.model.resume.ResumeList;
+
+public interface Person {
+    PersonName getName();
+    AchievementList getAchievements();
+    ResumeList getResumes();
+    int getNumAchievements();
+    void addAchievement(Achievement achievement);
+    void removeAchievement(Achievement achievement);
+    void addResume(Resume resume);
+    void removeResume(Resume resume);
+}
