+++ backend/model/resume/Resume.java
@@ -1,0 +1,6 @@
+package backend.model.resume;
+
+public interface Resume {
+    ResumeName getName();
+    ResumeContents getContents();
+}
