+++ backend/util/Message.java
@@ -1,0 +1,17 @@
+package backend.util;
+
+public enum Message {
+
+    INVALID_ACHIEVEMENT_INDEX_MESSAGE("Invalid achievement index received!");
+
+    private final String content;
+
+    private Message(String content) {
+        this.content = content;
+    }
+
+    public String getMessage() {
+        return content;
+    }
+
+}
