--- production/exception/ResumeNotFoundException.java
@@ -1,12 +1,0 @@
-package production.exception;
-
-public class ResumeNotFoundException extends BackendException {
-
-    private static final String RESUME_NOT_FOUND_MESSAGE =
-            "Could not find resume!";
-
-    public ResumeNotFoundException() {
-        super(RESUME_NOT_FOUND_MESSAGE);
-    }
-
-}
