--- production/model/achievement/Achievement.java
@@ -1,6 +1,0 @@
-package production.model.achievement;
-
-public interface Achievement {
-    AchievementName getName();
-    AchievementContents getContents();
-}
