--- production/model/resume/Resume.java
@@ -1,6 +1,0 @@
-package production.model.resume;
-
-public interface Resume {
-    ResumeName getName();
-    ResumeContents getContents();
-}
