--- production/model/resume/ResumeList.java
@@ -1,8 +1,0 @@
-package production.model.resume;
-
-public interface ResumeList {
-    boolean isEmpty();
-    boolean contains(Resume resume);
-    void add(Resume resume);
-    void remove(Resume resume);
-}
