--- production/util/ArgumentChecker.java
@@ -1,27 +1,0 @@
-package production.util;
-
-public class ArgumentChecker {
-
-    /**
-     * Checks that {@code condition} is true. Used for validating arguments to methods.
-     *
-     * @throws IllegalArgumentException if {@code condition} is false.
-     */
-    public static void checkArgument(Boolean condition) {
-        if (!condition) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    /**
-     * Checks that {@code condition} is true. Used for validating arguments to methods.
-     *
-     * @throws IllegalArgumentException with {@code errorMessage} if {@code condition} is false.
-     */
-    public static void checkArgument(Boolean condition, String errorMessage) {
-        if (!condition) {
-            throw new IllegalArgumentException(errorMessage);
-        }
-    }
-
-}
